+++ services/workflows-service/src/common/guards/session-auth.guard.ts
@@ -33,6 +33,7 @@ export class SessionAuthGuard implements CanActivate {
         user: authenticatedEntity?.user,
         type: 'user',
       });
+
       return true;
     }