+++ services/workflows-service/src/document/document.module.ts
@@ -1,10 +1,11 @@
-import { Module } from '@nestjs/common';
+import { forwardRef, Module } from '@nestjs/common';
 import { DocumentService } from './document.service';
 import { DocumentRepository } from './document.repository';
 import { DocumentControllerExternal } from './document.controller.external';
 import { PrismaModule } from '@/prisma/prisma.module';
 import { DocumentFileModule } from '@/document-file/document-file.module';
 import { FileModule } from '@/providers/file/file.module';
+// eslint-disable-next-line import/no-cycle
 import { WorkflowModule } from '@/workflow/workflow.module';
 import { UiDefinitionModule } from '@/ui-definition/ui-definition.module';
 import { WorkflowDefinitionModule } from '@/workflow-defintion/workflow-definition.module';
@@ -15,7 +16,7 @@ import { ProjectScopeService } from '@/project/project-scope.service';
     PrismaModule,
     DocumentFileModule,
     FileModule,
-    WorkflowModule,
+    forwardRef(() => WorkflowModule),
     UiDefinitionModule,
     WorkflowDefinitionModule,
   ],