+++ apps/backoffice-v2/CHANGELOG.md
@@ -12,6 +12,9 @@
 
 ### Patch Changes
 
+- fed1494bf4a7dde81915d4593765e7d89104c164: Bump
+- Updated dependencies [fed1494bf4a7dde81915d4593765e7d89104c164]
+  - @ballerine/common@0.9.96
 - version bump
 - Updated dependencies
   - @ballerine/blocks@0.2.43