+++ services/workflows-service/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Patch Changes
 
+- fed1494bf4a7dde81915d4593765e7d89104c164: Bump
+- Updated dependencies [fed1494bf4a7dde81915d4593765e7d89104c164]
 - version bump
 - Updated dependencies
   - @ballerine/common@0.9.96